pyproject.toml
[project]
name = "fastapibackend"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.11.13"
dependencies = [
    "aiofiles>=23.0.0",
    "aiohttp>=3.12.14",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "azure-storage-blob>=12.19.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.3",
    "sqlalchemy-utils>=0.41.2",
]
prod = [
    "gunicorn>=23.0.0",
]

[tool.ruff]
# Apply to all Python files including tests
extend-include = ["*.py"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# Basic rule sets
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ASYNC",# async/await best practices (perfect for FastAPI)
    "S",    # Security issues (important for auth/backend)
    "N",    # Naming conventions
]

ignore = [
    "B008",  # Allow function calls in argument defaults
    "N805",  # Allow non-self first argument (for Pydantic validators)
]

# Common ignores for test files
[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",    # Allow assert statements in tests
    "PLR2004", # Allow magic values in tests
    "S106",    # Allow hardcoded passwords in tests
    "ARG001",  # Allow unused function arguments (fixtures)
    "FBT001",  # Allow boolean positional arguments
    "FBT002",  # Allow boolean default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["fastapibackend"]

[tool.ruff.format]
# Format all files including tests
quote-style = "double"
indent-style = "space"